+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 * [GS-3051] Fix check certificate creating federated users
 * Kubernetes support
 * [EOS-4775] Fix step to remove user/group from tenant
+* [SPK-1168] Get log from mesos if CCT doesn't return log path
 
 ## 0.20.0 (December 17, 2020)